Removed sidebar from Data Page and created a sidebar component
